pyproject.toml
[tool.poetry]
name = "matriisi"
version = "0.7.0"
description = "A Trio library for writing Matrix clients/bots."
authors = ["Lura Skye <l@veriny.tf>"]
license = "GPL-3.0-or-later"
readme = "README.rst"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
trio = "^0.19.0"
httpx = "^0.19.0"
attrs = "^21.2.0"
cattrs = "^1.8.0"
h2 = "^4.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.2"
black = "^21.9b0"
prettyprinter = "^0.18.0"
coloredlogs = "^15.0.1"

[tool.isort]
profile = "black"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
